Some fixes to the keyboard handler:
------- WGL -------
* Keyboard handler now returns Unicode chars
* Fix numlock key interpreted as pause key
* Fix numlock, scrollock and capslock binds not working
* When numlock is off numeric keypad keys now return K_END, K_DOWNARROW,
* etc... Previously they always returned K_KP_<num>
regardless of numlock state

------- SDL -------
* Simulate press and release events when numlock and capslock get
* pressed, so that they can work similarly to the other keys
(excluding key repetition, it doesn't and will never work). It's a
workaround for the default SDL behaviour with these keys:
it generate a press event when a key gets enabled, a release event when
the key gets disabled. Scrollock already works as a
normal key.
* When numlock is off numeric keypad keys now return K_END, K_DOWNARROW,
* etc... Previously they always returned K_KP_<num>
regardless of numlock state

------- GLX -------
* Add numlock, scrollock and capslock keys

From: terencehill <piuntn@gmail.com>

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12085 d7cf8633-e32d-0410-b094-e92efae38249