NudgeOutOfSolid: fix "random" failures on slightly inclined planes
History:
This epsilon value was first added to the nudge distance in
94f4d6730223839ebd4cc08b71e5e625afb12d09
and was replaced with bbox expansion in
a319ca951a062e681a705952d25cb25dc118ae90

With the expansion approach we need to compare to separation to make
"good location" detection reliable, partly to avoid float precision
problems, partly to avoid compatibility problems with normal TraceBox.
For example: we drop an entity to floor successfully with TraceBox, but
then NudgeOutOfSolid decides it's in solid.

Also updated the startsolid conditions to suit the 2-pass approach.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>