sv_cheats: fix two segfaults when setting it to 0
Assuming the SVQC program is always running crashes when it's not.

Testing if the edict pointer is NULL isn't a valid way to stop at the
end of the client slots, which caused this to crash when Xonotic SVQC
was running.

See 20aa18d6b5d917caa92d4dde1973d907e3eb0c38

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>