PRVM: fix incorrect tempstring length in VM_tokenizebyseparator()
This caused memcpy() in PRVM_SetTempString() to copy too many bytes,
when the source string had the maximum length it could read past the end
and trigger a segfault.
Bug was introduced in bf5b4ab278c9423c0aca6f0b7ee6da72cf7a88a2 and looks
to be specific to that builtin.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>