host: make some startup errors non-fatal, related polishing and cleanup
It's annoying to abort DP just because of a bad menu progs or demo,
especially on Windows where accessing stdout is difficult so console
access is more important.

Shows a correct fallback message when the menu progs crashes (it's not
because of missing files).
Makes the missing files message more concise.

Performs a full cleanup after any QC crash, most importantly
progs->loaded is now false until that VM is restarted successfully.

Prevents menu failure interfering with more important VMs.

Removes the weird "early frame abort" jump point which is redundant now
because Host_Error, now the only code that ever aborts a frame, calls
Sys_Error instead of Host_AbortCurrentFrame during init.

Changes the framecount of the first frame from 0 to 1 so that when it's
>0 that means frame abort is now possible and safe.

Removes an old FIXME that was addressed.

Adds names for the host states.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>